requirements.txt
asgiref==3.5.2
bz2file==0.98
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
contourpy==1.0.7
Django==4.0.6
django-rest-framework==0.1.0
djangorestframework==3.14.0
matplotlib==3.6.3
mysql-connector-python==8.0.28
packaging==23.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
regex==2022.7.9
requests==2.28.1
sqlparse==0.4.2
threadpoolctl==3.1.0
tzdata==2022.1
urllib3==1.26.10
